+++ NEWS.md
@@ -1,3 +1,11 @@
+# dendroTools 1.0.5.
+* new function: monthly_response_seascorr()
+* titles of plot_extreme, plot_specific and plot_heatmap are updated
+* for daily_response() and monthly_response(), it is now possible to define method for correlation coefficient: "pearson", "kendall", "spearman". To do so, use , cor_method argument
+* daily_response output list is class "dmrs"
+* Using newly defined class "dmrs", new summary function is defined
+* Package version is changed to 1.0.5
+
 # dendroTools 1.0.4.
 * New data is available for examples demonstration: swit272_monthly_temperatures
 * New function is available: daily_response_seascorr(). This function implements partial correlations in the analysis of daily response functions